chore(coding-agent): simplify socket normalization coverage
