+++ packages/coding-agent/test/daemon-socket.test.ts
@@ -14,28 +14,9 @@ import {
 } from "../src/modes/daemon/daemon-socket.js";
 
 describe("normalizeSocketPath", () => {
-	it("collapses duplicate slashes", () => {
-		if (process.platform === "win32") {
-			return;
-		}
-
-		expect(normalizeSocketPath("/a//b.sock")).toBe("/a/b.sock");
-	});
-
-	it("strips trailing slashes", () => {
-		if (process.platform === "win32") {
-			return;
-		}
-
-		expect(normalizeSocketPath("/a/b.sock/")).toBe("/a/b.sock");
-	});
-
-	it("leaves canonical paths unchanged", () => {
-		if (process.platform === "win32") {
-			return;
-		}
-
-		expect(normalizeSocketPath("/a/b.sock")).toBe("/a/b.sock");
+	it("normalizes equivalent Unix spellings", () => {
+		if (process.platform === "win32") return;
+		expect(normalizeSocketPath("/a//b.sock/")).toBe("/a/b.sock");
 	});
 });
